+++ src/Component/OpenApi3/Tests/fixtures/issue-445/expected/Normalizer/IndexFieldNormalizer.php
@@ -38,7 +38,7 @@ class IndexFieldNormalizer implements DenormalizerInterface, NormalizerInterface
             return new Reference($data['$recursiveRef'], $context['document-origin']);
         }
         if (\array_key_exists('boost', $data) && \is_int($data['boost'])) {
-            $data['boost'] = (double) $data['boost'];
+            $data['boost'] = (float) $data['boost'];
         }
         if (\array_key_exists('ignoreForSearch', $data) && \is_int($data['ignoreForSearch'])) {
             $data['ignoreForSearch'] = (bool) $data['ignoreForSearch'];
